requirements.txt
bigquery==0.0.41
black==24.4.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dacktool==0.0.7
db-dtypes==1.2.0
dbstream==0.1.25
google-api-core==2.19.0
google-api-python-client==1.7.11
google-auth==2.29.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.23.1
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.64.0
grpcio-status==1.62.2
gspread==6.0.2
gspread_asyncio==2.0.0
httplib2==0.22.0
idna==3.7
libcst==1.3.1
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pandas==1.3.4
pathspec==0.12.1
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tzdata==2024.1
uritemplate==3.0.1
urllib3==2.2.1
